Fix 64 bits issue
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@746 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/internal_component/vci_mwmr_controller/tlmt/source/src/vci_mwmr_controller.cpp
@@ -241,7 +241,7 @@ namespace soclib { namespace tlmt {
 				  const tlmt_core::tlmt_time &time,
 				  void *private_data)
   {
-    int index = (int)private_data;
+	  int index = (int)(long)private_data;
 
     //Update the time local
     if(m_read_fifo[index].time < time)
@@ -288,7 +288,7 @@ namespace soclib { namespace tlmt {
 				   const tlmt_core::tlmt_time &time,
 				   void *private_data)
   {
-    int index = (int)private_data;
+	  int index = (int)(long)private_data;
 
     //Update the time local
     if(m_write_fifo[index].time < time)
